+++ package.json
@@ -14,6 +14,7 @@
   "license": "ISC",
   "homepage": "https://github.com/hueniverse/hueniversity",
   "dependencies": {
-    "hapi": "8.3.x"
+    "hapi": "8.3.x",
+    "hoek": "2.x.x"
   }
 }
